requirements.txt
moviepy==1.0.3
jsonpath-rw==1.4.0
pyttsx3==2.90
Pillow==7.1.2
